5908 Convert to QueryWebSocketST
